// src/components/organism/OurPartners/components/PartnersList/PartnersList.type.ts
export interface vendorType {
  id: string;
  rating?: number;
  numberOfStaff: number;
  logo?: string;
  cityName: string;
  companyNameAr: string;
  companyName: string;
  type: string;
  typeAr: string;
  location: string;
  locationAr: string;
  completedProjects: number;
  contactPhoneNumber: string;
  established: string;
  ongoingProjects: number;
  is_alphabetical?: boolean;
}